+++ src/App.tsx
@@ -1,23 +1,30 @@
 // import { useState } from "react";
-import "./App.css";
-import LoginPage from "./LoginPage";
-import HomePage from "./HomePage";
-import RegisterPage from "./RegisterPage";
+import "@/App.css";
+import { MainLayout } from "./pages/main-layout/MainLayout";
+import LoginPage from "./pages/login-page/LoginPage";
+import HomePage from "./pages/home-page/HomePage";
+import RegisterPage from "./pages/register-page/RegisterPage";
 import { Route, Routes } from "react-router-dom";
-import { Header } from "./components/layout/Header";
 
 function App() {
   return (
-    <>
-      <div className="App">
-        <Header />
-        <Routes>
-          <Route path="/home" element={<HomePage />} />
-          <Route path="/register" element={<RegisterPage />} />
-          <Route path="/login" element={<LoginPage />} />
-        </Routes>
-      </div>
-    </>
+    <div className="App">
+      <Routes>
+        {/* 不需要 Header 的頁面 */}
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/register" element={<RegisterPage />} />
+
+        {/* 有 Header 的頁面 (Header 另放在 MainLayout 之中) */}
+        <Route
+          path="/home"
+          element={
+            <MainLayout>
+              <HomePage />
+            </MainLayout>
+          }
+        />
+      </Routes>
+    </div>
   );
 }
